requirements.txt
scapy>=2.5.0
netaddr>=0.8.0
aiohttp>=3.8.0
orjson>=3.8.0
tqdm>=4.65.0
psutil>=5.9.0
pyyaml>=6.0.0
pandas>=1.5.0
plotly>=5.13.0
networkx>=3.0
numpy>=1.23.0
scikit-learn>=1.0.2
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
isort>=5.10.0
mkdocs>=1.4.0
mkdocs-material>=9.0.0
mkdocstrings>=0.20.0